+++ src/vrecip.h
@@ -14,7 +14,7 @@ struct _iproc_vrecip {
 	struct refcount refcount;
 };
 
-iproc_vrecip *iproc_vrecip_new(double *intvls, int64_t n);
+iproc_vrecip *iproc_vrecip_new(double *intvls, ssize_t n);
 iproc_vrecip *iproc_vrecip_ref(iproc_vrecip * v);
 void iproc_vrecip_unref(iproc_vrecip * v);
